Crate esp_vision

Example applications that demonstrate how to use OpenCV and ESP.

OpenCV is wrapped by a Rust library (currently a private repository)

ESP: Example-based Sensor Predictions

This application runs OpenCV CAMShift algorithm. Users select a region to track and the centroid of the tracked region is sent over to ESP (through a TCP connection).

Getting Started

Below is a possible ESP example that you can use (note the TCP port 8001).

#include <ESP.h>

TcpInputStream stream(8001, 2);
GestureRecognitionPipeline pipeline;

int timeout = 500;      // milliseconds
double null_rej = 0.4;

void setup() {
 stream.setLabelsForAllDimensions({"x", "y", "z"});
 useInputStream(stream);

 DTW dtw(false, true, null_rej);
 dtw.enableTrimTrainingData(true, 0.1, 75);

 pipeline.setClassifier(dtw);
 pipeline.addPostProcessingModule(ClassLabelTimeoutFilter(timeout));
 usePipeline(pipeline);

 registerTuneable(
     null_rej, 0.1, 5.0, "Variability",
     "How different from the training data a new gesture can be and "
     "still be considered the same gesture. The higher the number, the "
     "more different it can be.",
     [](double new_null_rej) {
         pipeline.getClassifier()->setNullRejectionCoeff(new_null_rej);
         pipeline.getClassifier()->recomputeNullRejectionThresholds();
     });

 registerTuneable(
     timeout, 1, 3000, "Timeout",
     "How long (in milliseconds) to wait after recognizing a "
     "gesture before recognizing another one.",
     [](double new_timeout) {
         ClassLabelTimeoutFilter* filter =
             dynamic_cast<ClassLabelTimeoutFilter*>(
                 pipeline.getPostProcessingModule(0));
         assert(filter != nullptr);
         filter->setTimeoutDuration(new_timeout);
     });
}

After running the ESP example, in this application, type cargo run would bring up the application.

Enjoy watching yourself :)

Reexports

extern crate std as std;

extern crate rust_vision;

use std::prelude::v1::*;

use rust_vision::*;

use std::io::prelude::*;

use std::net::TcpStream;

Structs

SelectionStatus
SelectionStatus tracks the region that users have selected for tracking.

Functions

main
The entry point to the application. Click into source for more information.

on_mouse
Mouse callback function. This gets called whenever a mouse event happens. Specifically in the implementation here we are populating the SelectionStatus struct so that CAMShift will track the right region.